-fixed a off by one bug with my getVisitCountForZone code -fixed how I'm clearing the database using an InMemory/testing store -adjusted the upper bounds for Max monsters and max zone choices -adjusted the calculateLvl method in CommonUtilities -wrote some more unit tests for ZoneHelper, only need to write one for GetZone and then I can move on -added a unit test class for Common Utilities

+++ HabitRPG2/CoreDataStackController.m
@@ -187,15 +187,26 @@
 
     -(void)deleteAllRecords{
         //this should only be called in testing situations.
-        [self deleteAllForEntity:HERO_ENTITY_NAME];
-        [self deleteAllForEntity:DATA_INFO_ENTITY_NAME];
-        [self deleteAllForEntity:MONSTER_ENTITY_NAME];
-        [self deleteAllForEntity:ZONE_ENTITY_NAME];
-        [self deleteAllForEntity:SETTINGS_ENTITY_NAME];
-        [self deleteAllForEntity:DAILY_ENTITY_NAME];
-        [self deleteAllForEntity:HABIT_ENTITY_NAME];
-        [self deleteAllForEntity:TODO_ENTITY_NAME];
-        [self deleteAllForEntity:GOOD_ENTITY_NAME];
+        NSArray<NSPersistentStore *> *stores = self.coordinator.persistentStores;
+        NSPersistentStore *ps = stores[0];
+        if([ps.type isEqualToString:@"InMemory"]){
+            NSError *err = nil;
+            [self.coordinator removePersistentStore:ps error:&err];
+            [self initializeCoreData];
+        }
+        else{
+            [self deleteAllForEntity:HERO_ENTITY_NAME];
+            [self deleteAllForEntity:DATA_INFO_ENTITY_NAME];
+            [self deleteAllForEntity:MONSTER_ENTITY_NAME];
+            [self deleteAllForEntity:ZONE_ENTITY_NAME];
+            [self deleteAllForEntity:SETTINGS_ENTITY_NAME];
+            [self deleteAllForEntity:DAILY_ENTITY_NAME];
+            [self deleteAllForEntity:HABIT_ENTITY_NAME];
+            [self deleteAllForEntity:TODO_ENTITY_NAME];
+            [self deleteAllForEntity:GOOD_ENTITY_NAME];
+            [self deleteAllForEntity:GOOD_ENTITY_NAME];
+            [self deleteAllForEntity:SUFFIX_ENTITY_NAME];
+        }
     }
 
 @end
